requirements.txt
lark-parser>=0.12.0
regex>=2026.1.15
colorama>=0.4.6
.